avoid race condition which removes the valid locks
In the monitor we load all processes and remove the process for which the
heartbeat is not set, then delete the lock for those invalid processes.

There is a race condition while deleting the invalid process that recently
started worker might get missed from all_process, which leads to deletion of its
lock. This will lead not respecting the limit since there is no lock for a
running worker.

The changes are taken from PR https://github.com/apolloio/sidekiq-limit_fetch/pull/1/files